not integration test but scenario test
