--- tests/utils/integration/types.ts
@@ -1,7 +1,0 @@
-export interface TestFixtures<TestCase extends BaseTestCase = BaseTestCase> {
-  testCases: TestCase[];
-}
-
-export interface BaseTestCase {
-  testName: string;
-}
