+++ tests/utils/scenario/types.ts
@@ -1,0 +1,7 @@
+export interface TestFixtures<TestScenario extends BaseTestScenario = BaseTestScenario> {
+  testScenarios: TestScenario[];
+}
+
+export interface BaseTestScenario {
+  scenarioName: string;
+}
